* commit '4c4301df0dbcb2830007fba3b319f1cadf17690d':
  Merge branch 'feature/FundsToLoan' into feature/fsl
  提交
  提交

+++ src/main/java/com/gqhmt/fss/architect/trade/service/FssTradeRecordService.java
@@ -67,7 +67,7 @@ public class FssTradeRecordService {
 	 */
 	public BigDecimal  getBankCode(String accNo) throws FssException{
 		FssAccountEntity fssAccountByAccNo = fssAccountService.getFssAccountByAccNo(accNo);
-		List<BankCardInfoEntity> queryInvestmentByCustId = bankCardInfoService.queryInvestmentByCustId(fssAccountByAccNo.getCustId());
+		List<BankCardInfoEntity> queryInvestmentByCustId = bankCardInfoService.queryInvestmentByCustId(fssAccountByAccNo.getCustId().intValue());
 		Application instance = Application.getInstance();
 		BigDecimal bankDealamountLimit = instance.getBankDealamountLimit(queryInvestmentByCustId.get(0).getParentBankId());
 		return bankDealamountLimit;
